cleaner

+++ h5p_extract.sh
@@ -7,7 +7,14 @@ mv $1 ${1%.h5p}.zip
 unzip ${1%.h5p}.zip -d extracted_content
 
 # Copy the required files
-cp -r extracted_content/content/ ./content/
+cp -r extracted_content/content/ ./content
 
-# Remove the extracted folder
-rm -rf extracted_content
+# Check if "libraries" folder exists
+if [ -d "libraries" ]
+then
+    # Remove the extracted folder
+    rm -rf extracted_content
+else
+    # Rename the extracted folder to "libraries"
+    mv extracted_content libraries
+fi
